requirements.txt
discord.py>=2.5.0
sentence-transformers>=5.0.0
transformers>=4.54.0
torch>=2.7.0
torchvision>=0.22.0
python-dotenv>=1.0.0
